Check usage. Much cleaner
