+++ lib/validateShp.js
@@ -1,0 +1,3 @@
+module.exports = function(input){
+
+}
